Split subfeature config to own file.
